chore: bump golangci-lint, fix lint, from sylabs 3533

+++ cmd/internal/cli/keyserver.go
@@ -2,7 +2,7 @@
 //   Apptainer a Series of LF Projects LLC.
 //   For website terms of use, trademark policy, privacy policy and other
 //   project policies see https://lfprojects.org/policies
-// Copyright (c) 2019-2023, Sylabs Inc. All rights reserved.
+// Copyright (c) 2019-2025, Sylabs Inc. All rights reserved.
 // Copyright (c) 2020, Control Command Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
@@ -174,7 +174,7 @@ var KeyserverRemoveCmd = &cobra.Command{
 }
 
 func setKeyserver(_ *cobra.Command, _ []string) {
-	if uint32(os.Getuid()) != 0 {
+	if os.Getuid() != 0 {
 		sylog.Fatalf("Unable to modify keyserver configuration: not root user")
 	}
 }
