+++ internal/pkg/util/user/identity_unix.go
@@ -2,7 +2,7 @@
 //   Apptainer a Series of LF Projects LLC.
 //   For website terms of use, trademark policy, privacy policy and other
 //   project policies see https://lfprojects.org/policies
-// Copyright (c) 2018-2019, Sylabs Inc. All rights reserved.
+// Copyright (c) 2018-2025, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -70,7 +70,7 @@ func CurrentOriginal() (*User, error) {
 	if err != nil {
 		return nil, err
 	}
-	return GetPwUID(uint32(uid))
+	return GetPwUID(uid)
 }
 
 // SetCurrentOriginal sets the current original user, used to set the
